[NT-117041] Work around mTLS issue
During TLS handshake, readServerCertificate() reads what looks like
regional-cache certificate's subject information into
certificateRequestMessageTLS13.certificateAuthorities. It then tries to
match this value with the k6 configured tlsAuth[].cert's issuer bytes.
These do not match as the issuer is different than the subject. I do not
know if Jetty or Go is at fault.

In order to work around this issue, I overrode Runner.newVU() to simply,
explicitly, return the k6 configured client certificate if it is the
only certificate configured via tls.Config.GetClientCertificate.

I would need to learn more about TLS handshake to understand the root
cause of this issue, but this allows me to use k6 to load test
regional-cache in the meantime.

Also, needed to statically link the executable so it is stand alone.

References
* https://github.com/golang/go/blob/7677616a263e8ded606cc8297cb67ddc667a876e/src/crypto/tls/handshake_client_tls13.go#L435
* https://github.com/golang/go/blob/7677616a263e8ded606cc8297cb67ddc667a876e/src/crypto/tls/handshake_messages.go#L1059